Fix a small issue with simplifying unchecked_enum_data.
Only simplify (unchecked_enum_data (enum payload)) when they refer to
the same element.

This hits when I apply the patch in <rdar://problem/16805316> and fix
the bug that it exposed. I looked at a handful of hits in the stdlib and
they all seem to be cases where we constant fold a conditional
branch (but have not yet simplified the branch), and the now-unreachable
side has code that creates a nil optional and then attempts to unwrap
it.

Swift SVN r17371