[irgen] When emitting code for the JIT, do not emit available external witness tables.
Currently we have in a hack that available_external entries are emitted
shared. This does not jive well in the JIT when one is attempting to
merge an external witness table with the linkonce_odr internal
available_external witness table.

We are not doing any optimization at the JIT level so there is no reason
to emit available_external witness tables at all.

So lets not do that.

Swift SVN r17423