Reenable the GlobalOpt pass.
Despite my comment in r17554, the pass was still disabled because of a
potential regression. After enough inlining, _cocoaStringSubscript
addressor is hoisted outside the string comparison, which is of course
a loop.

The regression is hard to measure, ~0.5%, so it's been decided that
we're going to live with it, rather than doing something nasty like
recognizing certain variable names. The post-WWDC fix:
<rdar://problem/16836228> Add an @cold attribute to identify unlikely
code paths.

Swift SVN r17609