Import 'NSError**' in parameters as typealias 'NSErrorResult'.
Previously 'NSError**' was imported directly as
ObjCMutablePointer<NSError?>, which is verbose and quite ugly
in the synthesized interfaces and overrides.  This remapping
is semantically equivalent by syntactically prettier.

This name needs to go through API review.

The test uses the actual SDK; the clang-importer-sdk didn't work
because the dump of the Foundation module only shows the pieces
from the Swift files, and the Objective-C pieces are just shown
as re-exported modules.  Testing the actual SDK provides good
fidelity anyway.

Swift SVN r17707