Rename @availability platforms to match Build configurations.
Originally my plan was to match Clang's spelling of platforms,
but this naming seems a bit more natural to just say "iOS" instead
of "ios" and "iOS_AppExtension" instead of "ios_app_extension".
What pushed me this direction was looking at how build configurations
used the platform name, which looked really natural.

The downside is that this does not match with Clang.  Comments
welcome to bike shed this one, but a goal is that build configurations
and @availablity are consistent with these names.

Swift SVN r17768