NoReturn folding: Set unreachable uses in different basic blocks to undef
The pass assumed that it is safe to delete instructions following a noreturn
call. If the instruction has an outside block user this would cause an failure
because we were deleting an instruction with uses.

Having control flow with uses from instructions after a noreturn call can happen
if we inline a transparent function after the noreturn call.

rdar://16852358

Swift SVN r17813