Increase the number of bits in KindAndValue.
We use KindAndValue storage to encode an argument index as well as other fields. The index argument overflows with long array literals.

KindAndValue is defined like this:  typedef Fixnum<29> KindAndValueStorage;

We encode the three values in 'value' with a 16 bit offset for the first field and 8 bit offset for the second field . 29 - 16 - 8 = 5 bits. This means that we can only support 2^5 = 32 arguments.



Swift SVN r17815