[PrintAsObjC] Don't include the ObjC header module in the generated header.
Importing a header with -import-objc-header causes the Clang importer to
provide an extra module to represent the header's content, and this was
showing up as "@import __ObjC;" in the /generated/ header for the target.
We should just not print anything there and let users import what's
necessary.

<rdar://problem/16917113>

Swift SVN r18081