[sil-aa] Given a pair of values (a,b), only dump (b,a) if the aa result for (b,a) differs from (a,b).
This reduces the amount of test output from aa-dumper which makes it
easier to write tests. We do not lose any test coverage since our
checking of Pair #s implies that if we emit an additional such pair, the
pair #s output by the aa-dumper will differ causing the test to fail.

rdar://16651852

Swift SVN r18141