[sil-aa] Add predicates to SILType that will be used in TBAA so that TBAA does not need to touch the AST.
We should never touch the AST from SIL if we don't need to. The specific
added methods are:

1. SILType::isBuiltinNativeObjectType().
2. SILType::isBuiltinRawPointerType().
3. SILType::isBuiltinType().
4. SILType::isBuiltinIntegerType().
5. SILType::isBuiltinFloatType().
6. SILType::isGenericType(). This supersedes the hasUnboundGenerics
function in Local.h. After WWDC, we should remove that method but I
don't want to update random parts of the SILPasses.
7. SILType::aggregateContainsRecord. Enables easy queries to determine
if a SILType T1 contains a separate SILType T2 as a member.

rdar://16651852

Swift SVN r18147