[docs] Revise the Array bridge/cast story
This change makes the story more coherent and consistent, and most
importantly, verifiable.  I've pushed ContiguousArray and Slice out of
the bridge/cast picture in order to simplify the description of Array.
Writing the story for those other array types (someday) in terms of this
one should be straightforward.

Swift SVN r18229