[devirtualizer] Fix issues where the position of Self in arguments/substitutions were confused.
I added some helpers to ApplyInst that should hopefully linguistically eliminate
the issue by allowing users of the API to not need to remember that the self
substitution is first, but the self argument is last.

We should really just remove the dichotomy. But that is for after WWDC.

I also disabled devirtualization of inherited protocol conformances for
protocol_methods. This will be less likely to be used than specialized
protocol conformances protocol_method devirtualization (which is
currently).

<rdar://problem/16951124>

Swift SVN r18282