Convert specializer.swift => specializer.sil. NFC.
For a long time we have had specializer tests written in swift instead of SIL
due to parsing issues. In order to start finding the errors there in I am
starting to (when I have a moment) convert tests that work from *.swift => *.sil
so we can test the SILParser functionality and preserve good behavior in the
parser.

Swift SVN r18298