Fix <rdar://problem/16953061> String outlet crashing in silgen
When type checking a patternbindingdecl with an initializer, we check the
initializer expression, then apply the inferred type to the pattern.  This
works except that we get down to a NamedPattern, see that it has various
attributes on it (e.g. iboutlet, weak) that affect the type of the pattern,
and we weren't re-propaging it back out through the pattern.  Do that.



Swift SVN r18355