Fix <rdar://problem/16663515> IBOutlet not adjusting getter/setter when making a property implicit unchecked optional
When the IBOutlet processing logic smashes a VarDecl to an implicit optional, it also has
to update the TypeLoc of any computed getters to include this, otherwise the getter will end
up with a type different from the variable, leading to later hilarity.

This is all really gross, but the TypeLoc mascinations that sema uses are not a good idea to
try to undo right now.


Swift SVN r18370