Reject casts to an implicitly unwrapped optional type.
This bans the syntax

  x as T!

because the user probably wanted "(x as T)!", but the parsing makes
more sense as "x as (T!)". When complaining about this syntax, provide
notes with Fix-Its for both parenthesizations, and recover as if the
user took the first one, which is by far more common. 

Addresses <rdar://problem/16952768>.


Swift SVN r18435