[Verifier] Clean up the scope stack if not visiting a node's children.
This was triggered when verifying a decl that was deserialized from a
swiftmodule, but only because the decl was the Swift version of a decl
available in a generated Objective-C header read by the Clang importer.
No test case because it would be a negative test based on the compiler's
current internal logic.

Swift SVN r18441