[stdlib] CMutablePointer: manual bridging support
When manually writing overlays, it's not uncommon to need to present an
interface that allows nil or an inout object to be passed, and
CMutablePointer allows that, but up to now it's been hard to work with
CMutablePointer on the inside of the manually-created thunks in the
overlay.  These small tweaks make that a bit more tolerable.

Swift SVN r18510