[serialization] Type and value cross-references are mutually exclusive.
In Swift, a type and a value (such as a function) will never have the same
name, but we import C tag types (structs, unions, and enums) into the same
namespace as values, which can lead to ambiguity. If the user has managed
to make clear which one they want in source code, though, we should preserve
that information during serialization.

<rdar://problem/16979415>

Swift SVN r18536