[g-arc-opts] Add GMALLOC like tool for running leaks while using lit.
Its easy to tell if we over-release memory since we will crash. It is
more difficult to know if we are leaking memory. This will help us to
catch such cases.

A test will be incoming once I finish a patch that enables lit support
for this (I wanted to separate out those two commits).

Swift SVN r18677